Add -P per_drv and -x fio_param options
If -P 1 and more than one drive is specified then the the cmdline
options are applied to each drive. For example, if 8 drives are
specfied and -P 0 (the default) then all the drives will be
processed by one fio cmd (with the --numjobs --iodepth etc.
settings).
With -P 1 8 fio cmds will be started.
Also, to help with testing, a '-x " --somefio_opt1=x --opt2=y "'
option was added. This lets you add or override options on the fio
cmdline.