vdbench jobs per numa node
for vdbench to get max perf it seems (to me) that we need to
process all the drives on a numa node in one job. For instance, if
there are some nvme drives on node 0 and other nvme drives on node
1 then we do 1 job per node, pinning each vdbench job to its node.
Also irqbalance needs to be disabled.
With these changes and the cmdline for vdb like in run_multi.sh
then I can hit pretty near the manufacturer specifications for the
drive.